Split game code into ui.js and engine.js
This makes it very easy to embed the game into any web page. main.js is now
only 4 lines of glue code for the stand-alone demo page.
Furthermore, this will be a good foundation to extend what is now only one
single screen with a playable demo into the fully-featured app that we
eventually want to get to.