Intertwine obstacle updating with collision detection
Also, roll back all obstacles to the moment of collision if one happened.
The funny thing here is that I was mainly motivated to this by the fact that
the cursor appeared partially inside the obstacle on collision. However, this
patch does not solve this. This is because we only transform the corner
vertices of obstacles into the circular coordinaate system. The lines that
connect the vertices are not transformed. If they were, they would form a
circular arc. Collision detection works in such a way that it would detect
collisions with the arc, not the trapezoid that is actually rendered. By
inspecting the original game, however, it became apparent that the problem
isn't solved there, either. Therefore, I will leave things as-is.